requirements.txt
Django==2.*
django-rest-framework
django-phone-field
django-crispy-forms
IPython
psycopg2
Pillow